do not allow redis job reclaim by same owner
Running a nonblocking conductor or two conductors on the same host will
re-execute the same job multiple times with the current implementation
of 'claim' for the redis jobboard backend. This is different from the
ZooKeeper jobboard backend, there the same owner of a job is not allowed
to reclaim the job again. If the same owner is allowed to reclaim the
job again there can be no concurrent execution on the same owner because
all jobs will be re-claimed and re-executed by the same owner every pass
as long as it's on the jobboard.

To reproduce this behavior:
1. Use the redis jobboard backend
2. Create a flow with a task that sleeps 10 seconds in the execute method
3. Post that flow as a job
4. Run a nonblocking conductor
5. It will claim and execute the same job multiple times in a loop until the first worker is finished and consumes the job

After this change it will not re-execute the same job multiple times